Attempted graphql implementation
This could work, but seems over kill right now. The hacking around the
lack of namespaced keyword support is ugly and crops up
everywhere. Getting pagination for our sharing matrix would require
reformatting it to move of an id->activity format.

Going to switch to a path of less resistance and go with a simple get resource.