Custom Joplin es6 migration manager
Introduces a custom Joplin plugin for managing ElasticSearch 6, as the
available version does not support this version.

There is duplication between this plugin and other namespaces. This is
because it's envisioned that this code might be pushable back to the
project and so it needs to stand alone.

It hasn't really been fully tested out, but it does find migrations
and executes them, there might be some more tweaking to do as our
index evolves, but with this level we should be able to iterate.