Elasticsearch client apply-func complete (#7)
Adds the ability to apply a function to a document stored in
elasticsearch. I've decided not support the usual case of retrying on
concurrent update failures as we shouldn't encounter these, so it just
felt like uneeded complication.

The acceptance tests explore the problem and prove that an informative
exception will be produced. For use it will mean a consumer
shutdown. If these prove to occur more than expected, then the try
loop can be reintroduced without much bother.

Metadata update component disabled as it's not ready yet.